Inherit from master configuration unless specified in slave configuration.
Fixes #13.

Typically with slave databases, most of the settings are actually the exact same as the master, except for host (usually).

So this allows you to write:

```yml
production:
  database: production
  username: username
  password: password
  encoding: utf8
  adapter:  mysql
  host:     master1
  pool:     50
  slave:
    host: slave1
```

Instead of:

```yml
production:
  database: production
  username: username
  password: password
  encoding: utf8
  adapter:  mysql
  host:     master1
  pool:     50
  slave:
    database: production
    username: username
    password: password
    encoding: utf8
    adapter:  mysql
    host:     slave1
    pool:     50
```

Which I think everyone can agree is much more DRY.

Of course, if the user wants to change any values or specify a completely different setup, all they are required to do is specify them in the `slave` configuration and they will be used instead of master.

But, for the vast majority of us, this makes the `slave` configuration insanely clean, easy and beautiful.